Refactor flowam hooks
