+++ python/tk_multi_breakdown2/flowam/reference.py
@@ -9,7 +9,8 @@
 # not expressly granted therein are reserved by Autodesk, Inc.
 
 from __future__ import annotations
-from typing import Any, Optional
+
+from typing import TYPE_CHECKING, Any, Optional
 
 import sgtk
 from tank_vendor.flow_integration_sdk import (
@@ -21,51 +22,11 @@ from tank_vendor.flow_integration_sdk import (
     utils,
 )
 
-
-def get_published_file_type(asset: objects.FlowAsset) -> Optional[dict[str, Any]]:
-    """Return a PublishedFileType-shaped stub derived from the asset's type_ids.
-
-    Uses the first type_id in asset.type_ids and resolves its display name via
-    schema.get_schema_display_name. Returns None if no type is available.
-
-    Args:
-        asset: A Flow Asset object exposing a type_ids attribute (list of str).
-    Returns:
-        A dict shaped like a SG PublishedFileType entity, or None.
-    """
-    if not asset or not getattr(asset, "type_ids", None):
-        return None
-    type_id = asset.type_ids[0]
-    try:
-        display_name = schema.get_schema_display_name(type_id)
-    except exceptions.FlowError:
-        display_name = type_id
-    return {"type": "PublishedFileType", "id": None, "code": display_name}
+if TYPE_CHECKING:
+    from ..api import FileItem
 
 
-def get_dependencies() -> list[dependency.DependencyData]:
-    """Return the list of asset dependencies found in current scene.
-    List will contain DependencyData objects which contain pertinent info
-    about each asset dependency instance found.
-
-    The DependencyData node contains detailed information of the dependency such as
-        * node_handle (node name)
-        * node_type (e.g. "reference")
-        * file_path (local file path of dependency)
-
-    Only the top level asset dependencies will be returned.
-    """
-    # Introspect current scene and get dependency tree
-    # (this will include asset and local dependencies together)
-    dep_tree = sgtk.platform.current_engine().flow_host.get_dependency_tree()
-
-    # Filter out only the top-level asset (internal) dependencies
-    asset_deps = dep_tree.get_internal_dependencies(top_level=True)
-
-    return asset_deps
-
-
-def get_scene_objects_and_publishes(
+def get_scene_objects(
     bg_task_manager: Optional[Any],
 ) -> tuple[list[dict[str, Any]], Any]:
     """
@@ -77,8 +38,6 @@ def get_scene_objects_and_publishes(
     so the breakdown app can build its file item model and determine statuses.
 
     Args:
-        manager: The breakdown manager instance, used to get published file fields.
-        published_file_fields: List of PublishedFile field names to query from Shotgun.
         bg_task_manager: Background task manager for async operations. If None, execute synchronously.
     Returns:
         A tuple of (scene_objects, published_file_data) where:
@@ -87,7 +46,7 @@ def get_scene_objects_and_publishes(
               file paths to a stub PublishedFile dict populated with MEDM data.
     """
 
-    flow_dependencies = get_dependencies()
+    flow_dependencies = _get_dependencies()
 
     scene_objects = [
         {
@@ -169,7 +128,297 @@ def get_scene_objects_and_publishes(
     return scene_objects, build_published_file_stubs()
 
 
-def update_dependency(
+def _get_published_file_type(asset: objects.FlowAsset) -> Optional[dict[str, Any]]:
+    """Return a PublishedFileType-shaped stub derived from the asset's type_ids.
+
+    Uses the first type_id in asset.type_ids and resolves its display name via
+    schema.get_schema_display_name. Returns None if no type is available.
+
+    Args:
+        asset: A Flow Asset object exposing a type_ids attribute (list of str).
+    Returns:
+        A dict shaped like a SG PublishedFileType entity.
+    """
+    if not asset or not getattr(asset, "type_ids", None):
+        return None
+    type_id = asset.type_ids[0]
+    try:
+        display_name = schema.get_schema_display_name(type_id)
+    except exceptions.FlowError:
+        display_name = type_id
+    return {"type": "PublishedFileType", "id": None, "code": display_name}
+
+
+def _get_dependencies() -> list[dependency.DependencyData]:
+    """Return the list of asset dependencies found in current scene.
+    List will contain DependencyData objects which contain pertinent info
+    about each asset dependency instance found.
+
+    The DependencyData node contains detailed information of the dependency such as
+        * node_handle (node name)
+        * node_type (e.g. "reference")
+        * file_path (local file path of dependency)
+
+    Only the top level asset dependencies will be returned.
+    """
+    # Introspect current scene and get dependency tree
+    # (this will include asset and local dependencies together)
+    dep_tree = sgtk.platform.current_engine().flow_host.get_dependency_tree()
+
+    # Filter out only the top-level asset (internal) dependencies
+    asset_deps = dep_tree.get_internal_dependencies(top_level=True)
+
+    return asset_deps
+
+
+# ============================
+# Get revisions in FPT format
+# ============================
+def get_latest_revision(
+    item: FileItem,
+    bg_task_manager: Optional[Any] = None,
+) -> dict[str, Any] | "BackgroundTaskManager":
+    """
+    Get the latest published file (revision) for a single item using MEDM.
+
+    Args:
+        item: FileItem to get the latest revision for.
+        bg_task_manager: If provided, the query runs async and the task id
+            is returned. Otherwise executes synchronously.
+    Returns:
+        If async, the background task id. Otherwise a dict shaped like a
+        SG PublishedFile entity representing the latest revision.
+    """
+    _bundle = sgtk.platform.current_bundle()
+
+    def _fetch_latest():
+        item_data = item.sg_data or {}
+        asset_id = item_data.get("sg_flow_asset_id")
+        if not asset_id:
+            raise exceptions.FlowError("No asset ID found for item")
+
+        project = _bundle.context.project
+
+        try:
+            asset = objects.FlowAsset(asset_id)
+            latest_revision = asset.get_latest_revision()
+            if not latest_revision:
+                _bundle.logger.warning(f"No latest revision found for asset {asset_id}")
+                raise exceptions.FlowError(
+                    f"No latest revision found for asset {asset_id}"
+                )
+        except exceptions.FlowError as e:
+            _bundle.logger.error(
+                f"Failed to get latest revision for asset {asset_id}: {e}"
+            )
+            raise
+
+        published_file_type = _get_published_file_type(asset)
+
+        local_path = latest_revision.get_storage_component_path(
+            component_purpose=globals.SOURCE_PURPOSE
+        )
+
+        created_at = asset.created_at
+
+        return {
+            "type": "PublishedFile",
+            "id": None,
+            "project": project,
+            "entity": {
+                "type": "Asset",
+                "id": asset_id,
+                "name": asset.name,
+            },
+            "name": item_data.get("name"),
+            "task": None,
+            "task.Task.sg_status_list": "No Status",
+            "tags": "No Tags",
+            "published_file_type": published_file_type,
+            "published_file_type.PublishedFileType.code": (
+                published_file_type["code"] if published_file_type else None
+            ),
+            "path": {"local_path": local_path} if local_path else None,
+            "version_number": asset.version_number,
+            "created_at": created_at,
+            "created_by.HumanUser.name": asset.created_by,
+            "description": asset.description,
+            "sg_flow_revision_id": latest_revision.id,
+            "sg_flow_asset_id": asset_id,
+            "sg_flow_version_id": asset.version_id,
+        }
+
+    if bg_task_manager:
+        return bg_task_manager.add_task(_fetch_latest)
+    return _fetch_latest()
+
+
+def get_assets_for_items(
+    items: list[FileItem],
+    bg_task_manager: Optional[Any] = None,
+) -> list[dict[str, Any]] | "BackgroundTaskManager":
+    """
+    Get all published file revisions for the given items using FlowAM.
+
+    Queries FlowAM for all numbered versions of each item's asset, returning
+    them as PublishedFile-shaped dicts so the breakdown model can build its
+    version mapping and determine statuses.
+
+    Args:
+        items: List of FileItem objects to get published files for.
+        bg_task_manager: If provided, the query runs async and the task id
+            is returned. Otherwise executes synchronously.
+    Returns:
+        If async, the background task id. Otherwise a list of dicts shaped
+        like SG PublishedFile entities, sorted newest-first per asset.
+    """
+
+    _bundle = sgtk.platform.current_bundle()
+
+    def _fetch_all_versions():
+        project = _bundle.context.project
+        result = []
+        processed_assets = {}
+
+        for item in items:
+            item_data = item.sg_data or {}
+            asset_id = item_data.get("sg_flow_asset_id")
+            if not asset_id:
+                continue
+
+            name = item_data.get("name")
+
+            if asset_id not in processed_assets:
+                try:
+                    asset = objects.FlowAsset(asset_id)
+                    versions = list(asset.iterate_versions())
+                    processed_assets[asset_id] = (asset, versions)
+                except exceptions.FlowError:
+                    _bundle.logger.error(
+                        f"Failed to query versions for asset {asset_id}"
+                    )
+                    continue
+
+            asset, versions = processed_assets[asset_id]
+            published_file_type = _get_published_file_type(asset)
+
+            for version in versions:
+                revision = version.revision
+                local_path = revision.get_storage_component_path(
+                    component_purpose=globals.SOURCE_PURPOSE
+                )
+
+                created_at = version.created_at
+
+                thumbnail_path = None
+                try:
+                    thumbnail_path = revision.get_thumbnail_file()
+                except exceptions.FlowError:
+                    _bundle.logger.error(
+                        f"Failed to get thumbnail path for revision {revision.id}"
+                    )
+
+                result.append(
+                    {
+                        "type": "PublishedFile",
+                        "id": None,
+                        "project": project,
+                        "entity": {
+                            "type": "Asset",
+                            "id": asset_id,
+                            "name": asset.name,
+                        },
+                        "name": name,
+                        "task": None,
+                        "task.Task.sg_status_list": "No Status",
+                        "tags": "No Tags",
+                        "published_file_type": published_file_type,
+                        "published_file_type.PublishedFileType.code": (
+                            published_file_type["code"] if published_file_type else None
+                        ),
+                        "path": {"local_path": local_path} if local_path else None,
+                        "version_number": version.version_number,
+                        "created_at": created_at,
+                        "created_by.HumanUser.name": version.created_by,
+                        "description": revision.comment,
+                        "sg_flow_revision_id": revision.id,
+                        "sg_flow_asset_id": asset_id,
+                        "sg_flow_version_id": version.id,
+                        "sg_flow_thumbnail_path": thumbnail_path,
+                    }
+                )
+
+        return result
+
+    if bg_task_manager:
+        return bg_task_manager.add_task(_fetch_all_versions)
+    return _fetch_all_versions()
+
+
+# ============================
+# Update depdendencies methods
+# ============================
+def update_to_latest(items: list[FileItem]) -> list[FileItem]:
+    """Update the given items in the scene.
+
+    Args:
+        items: list of file item object to update
+    Returns:
+        list of file item that were updated
+    """
+    items_to_update = []
+    for file_item in items:
+        res = _update_dependency(
+            file_item.sg_data["sg_flow_revision_id"],
+            node_handle=file_item.node_name,
+        )
+        if res:
+            items_to_update.append(file_item)
+
+    return items_to_update
+
+
+def update_to_revision(
+    item: Optional[FileItem],
+    item_data: Optional[dict[str, Any]] = None,
+) -> bool:
+    """Update the item to a specific version.
+
+    Args:
+        item: Dictionary representation of the FileItem to update
+        item_data: Dictionary of Flow Production Tracking data representing the target
+                published file revision to update to.
+
+    Returns:
+        True if the item requires the data model to update, else False will not
+        trigger a model update.
+    """
+    _bundle = sgtk.platform.current_bundle()
+    # Validate item_data contains the required revision ID
+    if not item_data or not item_data.get("sg_flow_revision_id"):
+        _bundle.logger.warning(
+            "Cannot update to revision: item_data is missing or lacks sg_flow_revision_id"
+        )
+        return False
+
+    # Validate item structure
+    flowam_data = item.get("sg_data") if item else None
+    if not flowam_data or not flowam_data.get("sg_flow_revision_id"):
+        _bundle.logger.warning(
+            "Cannot update to revision: item sg_data is missing or lacks sg_flow_revision_id"
+        )
+        return False
+
+    do_update = _update_dependency(
+        revision_id=flowam_data["sg_flow_revision_id"],
+        new_revision_id=item_data["sg_flow_revision_id"],
+        node_handle=item.get("node_name"),
+    )
+
+    return do_update
+
+
+def _update_dependency(
     revision_id: str,
     new_revision_id: str | None = None,
     node_handle: str | None = None,
